Replace custom log implementation with tracing
The tracing library offers a rich set of features and covers all the
implemented features of the logging implementation currently in use.

This commit replaces the various wrappers for the logging service with
the tracing equivalent. It also configures tracing to emit logs to
stderr and to a rolling file (although I recommend we remove this or
make it Windows only as systemd will handle logging for us if we write
to stdout or stderr).

Tracing makes it easy to emit structured logs if we prefer, integrate
with OpenTelemetry, and more.